requirements.txt
annotated-types==0.7.0
anyio==4.10.0
audioread==3.0.1
av==15.1.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
coqpit==0.0.17
coverage==7.10.7
ctranslate2==4.6.0
decorator==5.2.1
distlib==0.4.0
fastapi==0.117.1
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
humanfriendly==10.0
identify==2.6.14
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
mypy==1.18.2
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numba==0.62.0
numpy==1.26.4
onnxruntime==1.23.0
openai-whisper==20250625
packaging==25.0
pandas==2.2.3
pandas-stubs==2.3.2.250926
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
pre_commit==4.3.0
protobuf==6.32.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
ruff==0.13.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.13.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.2.2
tqdm==4.67.1
transformers==4.56.2
types-pytz==2025.2.0.20250809
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
virtualenv==20.34.0
wheel==0.45.1
